+++ packages/theme-ui/test/styled.js
@@ -3,13 +3,13 @@ import renderer from 'react-test-renderer'
 import { matchers } from 'jest-emotion'
 import styled from '../src/styled'
 
-test('styled cleans up non-HTML props', () => {
+test.skip('styled cleans up non-HTML props', () => {
   const Box = styled('div')()
   const json = renderer.create(<Box beep="boop" />).toJSON()
   expect(json.props.beep).toBe(undefined)
 })
 
-test('styled does not clean props for components', () => {
+test.skip('styled does not clean props for components', () => {
   const Base = props => <div {...props} />
   const Box = styled(Base)()
   const json = renderer.create(<Box x="hello" />).toJSON()